core/__init__.py
from .exchange import *
from .node import *
from .producer import *
from .consumer import *
from .pipeline import Pipeline
from .backtest import *

from dotenv import load_dotenv
load_dotenv()  # take environment variables from .env.