requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Backtesting==0.3.1
bokeh==2.3.2
certifi==2020.12.5
chardet==4.0.0
dateparser==1.0.0
greenlet==1.1.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
python-binance==1.0.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.14
TA-Lib==0.4.19
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
websockets==9.0.1
yarl==1.6.3
